linksets metadata

+++ bin/secondary/cr-linksets.sh
@@ -207,28 +207,31 @@ if [[ `is-pwd-a.sh                                                            cr
                echo ":linkset_$ls sio:has-member <$uri> ."               >> automatic/$bubble.ttl
             done
          done
-
+ 
          baseURI=${CSV2RDF4LOD_BASE_URI_OVERRIDE:-$CSV2RDF4LOD_BASE_URI}
          url="${baseURI}/source/$sourceID/file/cr-full-dump/version/latest/conversion/$base.nt.gz"
          echo "source/$sourceID.nt.gz <- $url"
          curl -s $url > source/$sourceID.nt.gz
          if [[ -n "$baseURI" && "$dryrun" != "true" ]]; then
-            echo "@prefix void: <http://rdfs.org/ns/void#> ."             > automatic/vocabulary.ttl
-            for term in `p-and-c.sh source/$sourceID.nt.gz | sort -u`; do
-               if [[ "$term" =~ http* ]]; then
-                  if [[ ${term%#*} != $term ]]; then
-                     echo " void:vocabulary <${term%#*}#>"
-                     echo "<$baseURI/void> void:vocabulary <${term%#*}> ."  >> automatic/vocabulary.ttl # No trailing '#'
-                  elif [[ ${term%/*} != $term ]]; then     # http://www.w3.org/TR/2011/NOTE-void-20110303/#vocabularies
-                     echo " void:vocabulary <${term%/*}/>"
-                     echo "<$baseURI/void> void:vocabulary <${term%/*}/> ." >> automatic/vocabulary.ttl
+            echo automatic/vocabulary.ttl
+            if [[ "$dryrun" != "true" ]]; then
+               echo "@prefix void: <http://rdfs.org/ns/void#> ."                > automatic/vocabulary.ttl
+               for term in `p-and-c.sh source/$sourceID.nt.gz | sort -u`; do
+                  if [[ "$term" =~ http* ]]; then
+                     if [[ ${term%#*} != $term ]]; then
+                        echo " void:vocabulary <${term%#*}#>"
+                        echo "<$baseURI/void> void:vocabulary <${term%#*}> ."  >> automatic/vocabulary.ttl # No trailing '#'
+                     elif [[ ${term%/*} != $term ]]; then     # http://www.w3.org/TR/2011/NOTE-void-20110303/#vocabularies
+                        echo " void:vocabulary <${term%/*}/>"
+                        echo "<$baseURI/void> void:vocabulary <${term%/*}/> ." >> automatic/vocabulary.ttl
+                     else
+                        echo "WARNING: `basename $0`: could not determine namespace for $term"
+                     fi
                   else
-                     echo "WARNING: `basename $0`: could not determine namespace for $term"
+                     echo "WARNING: `basename $0`: skipping non-HTTP term $term"
                   fi
-               else
-                  echo "WARNING: `basename $0`: skipping non-HTTP term $term"
-               fi
-            done
+               done
+            fi
          else
             echo "automatic/vocabulary.ttl - skipping b/c base URI not set."
          fi
